requirements.txt
streamlit>=1.24.0
torch>=2.0.0
opencv-python>=4.7.0
numpy>=1.24.0
Pillow>=9.5.0
sentence-transformers>=2.2.2